use top bottom and middle commit symbols
I like to use ┌ ├ └ for the commits, because it gives
me lines to follow visually. With flog v1 this works:

Plug 'rbong/vim-flog', { 'branch': 'v1' }
Plug 'TamaMcGlinn/flog-forest', { 'branch': 'v1' }

But with v2 the ability to customize the forest used in the background
was lost, so I have re-added this by checking if the commit has child
commits already drawn, and whether it has children that will later be
drawn, and still falling back to the • if neither (disconnected
commits).

The change to the MergeStart line just makes it a little more strict,
as it must be followed by a non-space character such as ──╮.
Otherwise, it thinks each middle-commit line is actually a
mergestart line and those commit lines become all-white.